neo4j_and_llms/requirements_pip.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
langchain-openai==0.0.6
langchainhub==0.1.14
langsmith==0.1.5
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.17.0
numpy==1.26.4
openai==1.12.0
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.27
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
youtube-search==2.1.2
